services/security-dna/anomaly-detector/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
neo4j==5.18.0
requests==2.31.0
scikit-learn==1.4.1.post1
numpy==1.26.4
mlflow==2.11.1
pandas==2.2.1
